Module Kwdcmd

KwdCmd: Keywords to Write Command Lines

A lightweight and partial porcelain around Cmdliner. Cmdliner is a powerful library for composable command line interfaces, but it is relatively low level. Most users have to look up the documentation every time they want to use the combinators correctly.

This thin wrapper library is best thought of as an executable cookbook for common Cmdliner usage patterns, or as programmatic documentation encoded in the type level.

The recipes are divided between module namespaces and detailed with named function arguments.

Ideals

Remain true to the Cmdliners compositional principles (no side-effecting shortcuts or other chicanery).

Unfamiliar users should be able to write their CLIs within 10 minutes of reading the docs.

Familiar users should find everything they need to know to write clean, expressive CLIS's self-documented in the type signatures of the library's modules and functions.

Please open an issue or add a feature if you think the library could be improved to better meet these ideals.

Usage

TODO

Example

open Kwdcmd

(** Application configuration *)
type kind =
  | Bin
  | Lib

type config =
  { name : string
  ; kind : kind
  }

(** Application executor *)
let run : config -> unit = fun _config -> print_endline "TODO"

(** CLI entrypoint *)
let () =
  Exec.run ~name:"My application" ~version:"0.0.1" ~doc:"project generator"
  @@ let+ name =
       Required.pos
         "NAME"
         ~conv:Arg.string
         ~nth:0
         ~doc:"The name of the new project"
         ()
  and+ kind =
    Optional.(
      flag_choice
        ~default:Bin
        [ c ~name:"bin" Bin ~doc:"create an executable binary"
        ; c ~name:"lib" Lib ~doc:"create a library"
        ])
  in
  run { name; kind }

Binding operators

The beauty of Cmdliner lies in its its applicative and composable API. Using this API is made much cleaner by means of binding operators. Naturally, you are not bound to use these. See the example above for usage.

val let+ : 'a Cmdliner.Term.t -> ('a -> 'b) -> 'b Cmdliner.Term.t

(let+) is Term.(const f $ t)

val and+ : 'a Cmdliner.Term.t -> 'b Cmdliner.Term.t -> ('a * 'b) Cmdliner.Term.t

(and+) is Term.(const (fun x y -> (x, y)) $ a $ b)

Constructing terms

We build our CLIs by defining composable terms that transform CLI args into OCaml values. Kwdcmd breaks terms into 2 classes.

module Required : sig ... end

module Optional : sig ... end

type 'a cmd = 'a Cmdliner.Term.t * Cmdliner.Term.info

val cmd : ?⁠man:Cmdliner.Manpage.block list -> name:string -> doc:string -> 'a Cmdliner.Term.t -> 'a cmd

A subcommand

val help_cmd : ?⁠version:string -> ?⁠doc:string -> ?⁠sdocs:string -> ?⁠exits:Cmdliner.Term.exit_info list -> ?⁠man:Cmdliner.Manpage.block list -> string -> 'a Cmdliner.Term.t * Cmdliner.Term.info

A custom help command

module Exec : sig ... end

Construct CLI entrypoints

Re-exports from cmdliner

val const : 'a -> 'a Cmdliner.Term.t

const (v : 'a) is a 'a Term.t: i.e. is a term that evaluates to v

val ($) : ('a -> 'b) Cmdliner.Term.t -> 'a Cmdliner.Term.t -> 'b Cmdliner.Term.t

f_term $ a_term evalutes to f a: i.e., apply a functional term to argument terms

val unit : unit Cmdliner.Term.t

unit evalutes to (), i.e. it is the unit term. Given a function f : unit -> unit), you can let term = lift f $ unit to execuate f when term is evaluated.

module Arg = Cmdliner.Arg

module Term = Cmdliner.Term