Module Kwdcmd.Default_handler

Default exit and error handlers for program execution.

type 'err err = [> `Msg of string ] as 'err

See Exec_handler.err

val err_handler : [> `Msg of string ] -> unit

err_handler err is the default handler for program execution errors:

(`Msg msg) results in an exit code of 1 with the msg printed to stderr

Any other err produces a failure

val exit_handler : ('a, [> `Msg of string ]) Stdlib.result Cmdliner.Term.result -> unit

exit_hander result converts the Term.result from a CLI entrypoint into a suitable exit conditions. It is the default exit handler for the entrypoints in Exec.

Uncaught exceptions return code 3

Parse errors or term errors (`Error `Parse | `Error `Term) return code 2

Execution errors are handled by err_handler.