Module Kwdcmd.Required

Required terms

These must be suplied or the resulting program will fail with an error indicating the missing arguments.

val pos : string -> ?rev:bool -> conv:'a Cmdliner.Arg.conv -> nth:int -> ?docs:string -> ?env:Cmdliner.Cmd.Env.info -> ?doc:string -> unit -> 'a Cmdliner.Term.t

pos docv ~conv ~nth is a positional argument at the nth position, giving a value derived by conv and named dcov in the help page.

val pos_all : string -> ?default:'a list -> conv:'a Cmdliner.Arg.conv -> ?docs:string -> ?env:Cmdliner.Cmd.Env.info -> ?doc:string -> unit -> 'a list Cmdliner.Term.t

val pos_left : string -> ?rev:bool -> ?default:'a list -> nth:int -> conv:'a Cmdliner.Arg.conv -> ?docs:string -> ?env:Cmdliner.Cmd.Env.info -> ?doc:string -> unit -> 'a list Cmdliner.Term.t