Module Kwdcmd

Kwdcmd: Keywords to Write Command Lines

Kwdcmd is a lightweight and partial porcelain around Cmdliner.

Cmdliner is a powerful library for composable command line interfaces, but it is relatively low level. Use of the combinators without reference to the docs requires a deep and thorough knowledge of the library. Kwdcmd wants to be able to be used without having to consult the docs.

This thin wrapper library is best thought of as an executable cookbook for common Cmdliner usage patterns, or as programmatic documentation encoded in the type level.

The recipes are divided between module namespaces and detailed with named function arguments.

Ideals

Remain true to Cmdliners compositional principles (no side-effecting shortcuts or other chicanery).

Unfamiliar users should be able to write their CLIs within 10 minutes of reading the docs.

Familiar users should find everything they need to know to write clean, expressive CLIS's self-documented in the type signatures of the library's modules and functions.

Please open an issue or add a feature if you think the library could be improved to better meet these ideals.

Usage

The basic usage pattern is as follows:

Define Cmdliner terms, i.e., parsers producing values from CLI arguments, using the constructors.

Use the binding operators to declare a collection of term parsers which are then fed to a program.

Apply an executor to run the parser and the program.

CLI programs are defined either as a single command or with multiple subcommands. Here are two short examples to illustrate each case:

A single command parser

open Kwdcmd

(** Application configuration *)
type kind =
  | Bin
  | Lib

type config =
  { name : string
  ; kind : kind
  }

(** Application executor *)
let run : config -> unit = fun _config -> print_endline "TODO"

(** CLI entrypoint *)
let () =
  Exec.run ~name:"My application" ~version:"0.0.1" ~doc:"project generator"
  @@ let+ name =
       Required.pos
         "NAME"
         ~conv:Arg.string
         ~nth:0
         ~doc:"The name of the new project"
         ()
  and+ kind =
    Optional.(
      flag_choice
        ~default:Bin
        [ c ~name:"bin" Bin ~doc:"create an executable binary"
        ; c ~name:"lib" Lib ~doc:"create a library"
        ])
  in
  run { name; kind }

A sub-command parser

A sub-command interface to a simple utility for looking up emoji:

module Example_cli (Progn : sig
    val lookup_name : string -> (unit, _ err) cmd_result
    val lookup_unicode : string -> (unit, _ err) cmd_result
    val emojify : Fpath.t -> (unit, _ err) cmd_result
  end )
= struct
  open Kwdcmd

  let () =
    Exec.commands
      ~name:"emojitsu"
      ~version:"0.0.1"
      ~doc:"Techniques for dealing with emoji"
      [ ( cmd
            ~name:"find-name"
            ~doc:"Find the name of an emoji given its unicode"
          @@ let+ unicode = Required.pos "UNICODE" ~conv:Arg.string ~nth:0 () in
          Progn.lookup_name unicode )
      ; ( cmd
            ~name:"find-unicode"
            ~doc:"Find the unicode of an emoji given its name"
          @@ let+ name = Required.pos "EMOJI_NAME" ~conv:Arg.string ~nth:0 () in
          Progn.lookup_unicode name )
      ; ( cmd
            ~name:"emojify"
            ~doc:
              "Replace all names of the form :emoji_name: with the corresponding \
               unicode in the given file"
          @@ let+ name =
               Required.pos
                 "FILE"
                 ~conv:Arg.(conv (Fpath.of_string, Fpath.pp))
                 ~nth:0
                 ()
          in
          Progn.emojify name )
      ]
end

Binding operators

The beauty of Cmdliner lies in its its composable, applicative API. Use of this API is made cleaner by means of the binding operators. Naturally, you are not bound to use these.

The general schema is

let+ value_1 = term_1
and+ value_2 = term_2
(* ... *)
and+ value_n = term_n
in
program value_1 value_2 (* ... *) value_n

where each value_i will be the value obtained by parsing by the the CLI term specified in term_i.

See the example above for usage.

val let+ : 'a Cmdliner.Term.t -> ( 'a -> 'b ) -> 'b Cmdliner.Term.t

(let+) is Term.(const f $ t)

val and+ : 'a Cmdliner.Term.t -> 'b Cmdliner.Term.t -> ('a * 'b) Cmdliner.Term.t

(and+) is Term.(const (fun x y -> (x, y)) $ a $ b)

Constructing terms

We build our CLIs by defining composable terms that transform CLI args into OCaml values. Kwdcmd breaks terms into 2 classes.

module Required : sig ... end

module Optional : sig ... end

type 'a cmd = 'a Cmdliner.Term.t * Cmdliner.Cmd.info

The type of subcommands

val cmd : ?man:Cmdliner.Manpage.block list -> name:string -> doc:string -> 'a Cmdliner.Term.t -> 'a Cmdliner.Cmd.t

A subcommand

type 'err err = [> `Msg of string ] as 'err

Errors the program configured by the CLI can produce. This does not include errors resulting from parsing the CLI. Those are represented by Term.result.

type ('a, 'err) cmd_result = ( 'a, 'err err ) Stdlib.result

Executing CLIs

module type Exec_handler = sig ... end

module Default_handler : sig ... end

Default exit and error handlers for program execution.

module type Exec = sig ... end

module Exec : Exec

Re-exports from cmdliner

See the Cmdliner documentation for details.

val const : 'a -> 'a Cmdliner.Term.t

const v : 'a) is a 'a Term.t: i.e. is a term that evaluates to v

val ($) : ( 'a -> 'b ) Cmdliner.Term.t -> 'a Cmdliner.Term.t -> 'b Cmdliner.Term.t

f_term $ a_term evalutes to f a: i.e., apply a functional term to argument terms

val unit : unit Cmdliner.Term.t

unit evalutes to (), i.e. it is the unit term. Given a function f : unit -> unit), you can let term = lift f $ unit to execuate f when term is evaluated.

module Arg = Cmdliner.Arg

module Term = Cmdliner.Term

module Manpage = Cmdliner.Manpage