Module Kwdcmd.Optional

Optional terms

val values : string -> flags:string list -> ?⁠default:'a list -> conv:'a Cmdliner.Arg.conv -> ?⁠docs:string -> ?⁠env:Cmdliner.Arg.env -> ?⁠doc:string -> unit -> 'a list Cmdliner.Term.t

val value : string -> flags:string list -> default:'a -> conv:'a Cmdliner.Arg.conv -> ?⁠docs:string -> ?⁠env:Cmdliner.Arg.env -> ?⁠doc:string -> unit -> 'a Cmdliner.Term.t

val flag : flags:string list -> ?⁠docs:string -> ?⁠docv:string -> ?⁠env:Cmdliner.Arg.env -> ?⁠doc:string -> unit -> bool Cmdliner.Term.t

val all_from : string -> conv:'a Cmdliner.Arg.conv -> nth:int -> ?⁠default:'a list -> ?⁠docs:string -> ?⁠env:Cmdliner.Arg.env -> ?⁠doc:string -> unit -> 'a list Cmdliner.Term.t

All the values following from the nth position on

val pos : string -> conv:'a Cmdliner.Arg.conv -> nth:int -> ?⁠docs:string -> ?⁠env:Cmdliner.Arg.env -> ?⁠doc:string -> unit -> 'a option Cmdliner.Term.t

type 'a choice = 'a * Cmdliner.Arg.info

val c : doc:string -> name:string -> 'option -> 'option choice

A choice flag

val flag_choices : defaults:'option list -> 'option choice list -> 'option list Cmdliner.Term.t

Choose between any number of the given choice flags, can be repeated if none of the choices are provded, then use the default choices

val flag_choice : default:'option -> 'option choice list -> 'option Cmdliner.Term.t

Choose a single one of the given choice flags