Module Kwdcmd.Required

Required terms

These must be suplied or the resulting program will fail with error indicating the missing arguments.

val pos : string -> conv:'a Cmdliner.Arg.conv -> nth:int -> ?⁠docs:string -> ?⁠env:Cmdliner.Arg.env -> ?⁠doc:string -> unit -> 'a Cmdliner.Term.t

pos docv ~conv ~nth is a positional argument at the the nth position, giving a value derived by conv and named dcov in the help page.